Added third-party libraries gtk themes, which allows you to apply different themes to the interface, and ThemedTk, which allows you to create windows with the applied theme.
Removed the display of the window with a message about the correctness of the answer. Now, when you click the "Reply" button, the answer will be checked and the transition to the next card will be carried out without delay.

Added closing the testing window after completing all the cards and displaying a message with the test result indicating the number of correct answers.

Added shuffling of cards before testing using the random.shuffle() function. This prevents the same answer from appearing for each card.